[IMP] Mail issues Signature problem and Whole company, Company News and Company Jobs should not be deleted

+++ addons/mail/mail_group.py
@@ -171,14 +171,15 @@ class mail_group(osv.Model):
         # Cascade-delete mail aliases as well, as they should not exist without the mail group.
         mail_alias = self.pool.get('mail.alias')
         alias_ids = [group.alias_id.id for group in groups if group.alias_id]
-        # Delete mail_group
-        data_id = []
-        data_id.append(self.pool['ir.model.data'].get_object_reference(cr, uid, 'mail', 'group_all_employees')[1])
-        data_id.append(self.pool['ir.model.data'].get_object_reference(cr, uid, 'portal', 'company_news_feed')[1])
-        data_id.append(self.pool['ir.model.data'].get_object_reference(cr, uid, 'portal', 'company_jobs')[1])
-        for id in data_id:
-            if id in ids :
-                raise osv.except_osv(_('Warning!'), _('You can not delete this group because some other modules are depend on it')) 
+        # Delete 
+        mail_grp = self.pool.get('ir.model.data')
+        data_ids = []
+        data_ids.append(mail_grp.get_object_reference(cr, uid, 'mail', 'group_all_employees')[1])
+        data_ids.append(mail_grp.get_object_reference(cr, uid, 'portal', 'company_news_feed')[1])
+        data_ids.append(mail_grp.get_object_reference(cr, uid, 'portal', 'company_jobs')[1])
+        data_ids =  [id for id in data_ids if id in ids]
+        if data_ids:
+            raise osv.except_osv(_('Warning!'), _('You can not delete this group because some other modules are depend on it')) 
         res = super(mail_group, self).unlink(cr, uid, ids, context=context)
         # Delete alias
         mail_alias.unlink(cr, SUPERUSER_ID, alias_ids, context=context)
